Split Event code out into Sync
Split Event.IssueComment out into Sync.Comment

Split Event.Issues out into Sync.Issue

Move RepoFinder into Sync.Utils

Split Event.PullRequest out into Sync.PullRequest

Improve docs

Refactor tests for issues event and issue sync

Refactor tests for issues comment event and comment sync

Refactor tests for pull request event and pull request sync
