+++ lib/code_corps/github/sync/issue/issue.ex
@@ -1,15 +1,20 @@
-defmodule CodeCorps.GitHub.Sync.PullRequest do
+defmodule CodeCorps.GitHub.Sync.Issue do
   alias CodeCorps.{
+    GitHub,
     GitHub.Sync.Utils.RepoFinder,
+    GitHub.Sync.Issue.Task,
     Repo,
     Task
   }
-  alias CodeCorps.GitHub.Sync.PullRequest.GithubPullRequest, as: GithubPullRequestSyncer
-  alias CodeCorps.GitHub.Sync.PullRequest.Task, as: PullRequestTaskSyncer
-  alias CodeCorps.GitHub.Sync.User.RecordLinker, as: UserRecordLinker
+  alias GitHub.Sync.Issue.GithubIssue, as: IssueGithubIssueSyncer
+  alias GitHub.Sync.Issue.Task, as: IssueTaskSyncer
+  alias GitHub.Sync.User.RecordLinker, as: UserRecordLinker
   alias Ecto.Multi
 
   @type outcome :: {:ok, list(Task.t)}
+                 | {:error, :not_fully_implemented}
+                 | {:error, :unexpected_action}
+                 | {:error, :unexpected_payload}
                  | {:error, :repository_not_found}
                  | {:error, :validation_error_on_inserting_user}
                  | {:error, :multiple_github_users_matched_same_cc_user}
@@ -17,15 +22,14 @@ defmodule CodeCorps.GitHub.Sync.PullRequest do
                  | {:error, :unexpected_transaction_outcome}
 
   @doc ~S"""
-  Syncs a GitHub pull request API payload with our data.
+  Syncs a GitHub issue API payload with our data.
 
   The process is as follows:
 
-  - match payload with affected `CodeCorps.GithubRepo` record using
+  - match payload with `CodeCorps.GithubRepo` record using
     `CodeCorps.GitHub.Sync.Utils.RepoFinder`
-  - match with `CodeCorps.User` using
-    `CodeCorps.GitHub.Event.PullRequest.UserLinker`
-  - for each `CodeCorps.ProjectGithubRepo` belonging to matched repo:
+  - match with `CodeCorps.User` using `CodeCorps.GitHub.Sync.User.RecordLinker`
+  - for each `CodeCorps.ProjectGithubRepo` belonging to the matched repo:
     - create or update `CodeCorps.Task` for the `CodeCorps.Project`
 
   If the sync succeeds, it will return an `:ok` tuple with a list of created or
@@ -45,16 +49,15 @@ defmodule CodeCorps.GitHub.Sync.PullRequest do
   @spec operational_multi(map) :: Multi.t
   defp operational_multi(payload) do
     Multi.new
-    Multi.new
     |> Multi.run(:repo, fn _ -> RepoFinder.find_repo(payload) end)
-    |> Multi.run(:pull_request, fn %{repo: github_repo} -> link_pull_request(github_repo, payload) end)
-    |> Multi.run(:user, fn %{pull_request: github_pull_request} -> UserRecordLinker.link_to(github_pull_request, payload) end)
-    |> Multi.run(:tasks, fn %{pull_request: github_pull_request, user: user} -> github_pull_request |> PullRequestTaskSyncer.sync_all(user, payload) end)
+    |> Multi.run(:issue, fn %{repo: github_repo} -> link_issue(github_repo, payload) end)
+    |> Multi.run(:user, fn %{issue: github_issue} -> UserRecordLinker.link_to(github_issue, payload) end)
+    |> Multi.run(:tasks, fn %{issue: github_issue, user: user} -> github_issue |> IssueTaskSyncer.sync_all(user, payload) end)
   end
 
-  @spec link_pull_request(GithubRepo.t, map) :: {:ok, GithubIssue.t} | {:error, Ecto.Changeset.t}
-  defp link_pull_request(github_repo, %{"pull_request" => attrs}) do
-    GithubPullRequestSyncer.create_or_update_pull_request(github_repo, attrs)
+  @spec link_issue(GithubRepo.t, map) :: {:ok, GithubIssue.t} | {:error, Ecto.Changeset.t}
+  defp link_issue(github_repo, %{"issue" => attrs}) do
+    IssueGithubIssueSyncer.create_or_update_issue(github_repo, attrs)
   end
 
   @spec marshall_result(tuple) :: tuple
